Added onCollision event to locationComponent.

+++ src/drawingHandler.cpp
@@ -9,8 +9,10 @@ void DrawingHandler::draw(std::map<int, Entity*>& entityMap) {\
     for (it = entityMap.begin(); it != entityMap.end(); ++it) {
         Entity* entity = it->second;
         if (entity->art && entity->location) {
-            SDL_Rect rect = {entity->location->x, entity->location->y,
-                             entity->location->width, entity->location->height};
+            SDL_Rect rect = { (int) entity->location->x,
+                              (int) entity->location->y,
+                              entity->location->width,
+                              entity->location->height };
             SDL_RenderCopy(this->renderer, entity->art->texture, NULL, &rect);
         }
     }
